+++ release_notes.md
@@ -1,6 +1,7 @@
 ###In Development
   - [#257](https://github.com/Mehdik/Humanizer/pull/257): Added German localisation for ToOrdinalWords and Ordinalize
   - [#261](https://github.com/Mehdik/Humanizer/pull/261): Added future dates to Portuguese - Brazil
+  - [#268](https://github.com/Mehdik/Humanizer/pull/268): Added humanization of collections
 
 [Commits](https://github.com/MehdiK/Humanizer/compare/v1.25.4...master)